refactor(position): 디렉터리 위치 piece -> position 변경
