+++ src/main/java/domain/movestrategy/MoveStrategy.java
@@ -1,7 +1,7 @@
 package domain.movestrategy;
 
+import domain.board.Position;
 import domain.piece.Piece;
-import domain.piece.Position;
 import java.util.List;
 import java.util.Map;
