+++ src/test/java/domain/board/BoardTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.piece.Piece;
 import domain.piece.PieceType;
-import domain.piece.Position;
 import domain.piece.Team;
 import java.util.List;
 import java.util.Map;
