+++ src/test/java/domain/movestrategy/HorseMoveStrategyTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.piece.Piece;
 import domain.piece.PieceType;
-import domain.piece.Position;
+import domain.board.Position;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
